Removed Catcher as it generated unnecesairy issues
Catcher objects are supposed to be used to solve problems with buffering
of images in camera devices, when they are not read frequently enough.

Current form of catcher presents some issues connected with multithreading,
which solving is not main concern of this project (it stopped me for nearly
two months, I have enough).